Ensure that hash args get wrapped in an array.
To enable enumerations of arguments where the argument could be
a single scalar value instead of a list of a single scalar value,
we used to wrap all non-enumerable types in an array. This broke
hash arguments, that still needs to be wrapped. Instead wrap all
non-array-like (responds to to_ary).